rename(common): src내 디렉토리 추가 생성 및 파일 분류
