+++ src/App.js
@@ -1,7 +1,7 @@
-import GlobalStyle from './shared/GlobalStyle';
 import Routers from './Router';
 import Header from './components/shared/Header';
 import { IssueProvider } from './context/IssueContext';
+import GlobalStyle from './styles/GlobalStyle';
 
 const App = () => {
   return (
