+++ src/Router.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Detail from './pages/Detail';
-import List from './pages/List';
+import DetailPage from './pages/DetailPage';
+import ListPage from './pages/ListPage';
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={<List />} />
-        <Route path="/issue/:id" element={<Detail />} />
+        <Route path="/" element={<ListPage />} />
+        <Route path="/issue/:id" element={<DetailPage />} />
       </Routes>
     </BrowserRouter>
   );
