--- src/pages/Detail.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import DetailFrame from '../components/Detail/DetailFrame';
-
-const Detail = () => {
-  return <DetailFrame />;
-};
-
-export default Detail;
