+++ src/pages/DetailPage.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import DetailFrame from '../components/detail/DetailFrame';
+
+const DetailPage = () => {
+  return <DetailFrame />;
+};
+
+export default DetailPage;
