--- src/pages/List.jsx
@@ -1,7 +1,0 @@
-import ListFrame from '../components/List/ListFrame';
-
-const List = () => {
-  return <ListFrame />;
-};
-
-export default List;
