+++ src/pages/ListPage.jsx
@@ -1,0 +1,7 @@
+import ListFrame from '../components/list/ListFrame';
+
+const ListPage = () => {
+  return <ListFrame />;
+};
+
+export default ListPage;
